# Add filtering to performance comparisons