--- a/package.json
+++ b/package.json
@@ -30,6 +30,7 @@
     "fs-extra": "^0.26.2",
     "fs-generate": "^1.0.0",
     "glob": "^6.0.1",
+    "micromatch": "^2.3.5",
     "mocha": "^2.3.4",
     "readdirp": "^2.0.0",
     "sinon": "^1.17.2",