--- a/package.json
+++ b/package.json
@@ -29,7 +29,9 @@
     "chai": "^3.4.1",
     "fs-extra": "^0.26.2",
     "fs-generate": "^1.0.0",
+    "glob": "^6.0.1",
     "mocha": "^2.3.4",
+    "readdirp": "^2.0.0",
     "sinon": "^1.17.2",
     "sinon-chai": "^2.8.0"
   }