--- a/perf/tests.js
+++ b/perf/tests.js
@@ -1,5 +1,6 @@
 var sysPath = require('path');
 
 var DIR = sysPath.resolve(sysPath.join(__dirname, '..', 'node_modules'));
+var PATTERN = '**/*.js';
 
-require('./run-directory')(DIR);
+require('./run-directory')(DIR, PATTERN);