Don't know whether you'll want this in master, but here's a quick pass at adding some degree of utilizing glob filtering to the benchmarks.

These are the results I got:

```
// PATTERN = '**/*.js'

Comparing Walk walk-filtered/node_modules/**/*.js
Default options x 20.01 ops/sec ±1.54% (52 runs sampled)
Serial (fs) x 9.74 ops/sec ±0.62% (50 runs sampled)
Parallel (fs) x 20.36 ops/sec ±0.79% (57 runs sampled)
Parallel (gfs) x 20.31 ops/sec ±0.76% (52 runs sampled)
Parallel limit (fs, 10) x 19.63 ops/sec ±0.63% (50 runs sampled)
Parallel limit (fs, 50) x 20.49 ops/sec ±0.69% (53 runs sampled)
Parallel limit (fs, 100) x 20.63 ops/sec ±0.74% (54 runs sampled)
Parallel limit (gfs, 10) x 19.40 ops/sec ±0.61% (50 runs sampled)
Parallel limit (gfs, 50) x 20.25 ops/sec ±0.69% (52 runs sampled)
Parallel limit (gfs, 100) x 20.11 ops/sec ±0.69% (52 runs sampled)
readdirp x 18.76 ops/sec ±1.95% (60 runs sampled)
glob x 18.38 ops/sec ±1.55% (88 runs sampled)
Fastest is Parallel limit (fs, 100),Parallel limit (fs, 50)
```

```
// PATTERN = false

Comparing Walk walk-filtered/node_modules
Default options x 20.60 ops/sec ±1.81% (53 runs sampled)
Serial (fs) x 10.18 ops/sec ±0.46% (52 runs sampled)
Parallel (fs) x 21.11 ops/sec ±0.74% (54 runs sampled)
Parallel (gfs) x 20.76 ops/sec ±0.75% (53 runs sampled)
Parallel limit (fs, 10) x 20.33 ops/sec ±0.70% (52 runs sampled)
Parallel limit (fs, 50) x 20.95 ops/sec ±0.87% (54 runs sampled)
Parallel limit (fs, 100) x 21.16 ops/sec ±0.92% (54 runs sampled)
Parallel limit (gfs, 10) x 19.80 ops/sec ±0.62% (51 runs sampled)
Parallel limit (gfs, 50) x 20.76 ops/sec ±0.75% (53 runs sampled)
Parallel limit (gfs, 100) x 20.79 ops/sec ±0.59% (53 runs sampled)
readdirp x 19.41 ops/sec ±2.04% (51 runs sampled)
glob x 14.03 ops/sec ±1.19% (70 runs sampled)
Fastest is Parallel limit (fs, 100),Parallel (fs)
```